% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baltimoremls-package.R
\docType{package}
\name{baltimoremls-package}
\alias{baltimoremls}
\alias{baltimoremls-package}
\title{baltimoremls: Read and Format Baltimore MLS Data}
\description{
A package with helper functions for reading and formatting Bright MLS property sales data for Baltimore City provided to city agencies through a parntership with Live Baltimore.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/elipousson/baltimoremls/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
