% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_scale_mls_prices}
\alias{format_scale_mls_prices}
\title{Scale MLS price values by a supplied factor value}
\usage{
format_scale_mls_prices(
  data,
  ...,
  factor = 1000,
  cols = c("list_price", "close_price", "original_list_price", "sales_price")
)
}
\description{
Scale MLS price values by a supplied factor value
}
\keyword{internal}
