% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mls_data.R
\name{get_mls_data}
\alias{get_mls_data}
\title{Get Bright MLS Data for Baltimore City}
\usage{
get_mls_data(
  data,
  location = NULL,
  crop = TRUE,
  trim = FALSE,
  collapse_types = TRUE,
  ...,
  crs = 3857
)
}
\arguments{
\item{data}{File path or URL for Bright MLS Sales data.}

\item{location}{sf object. If multiple areas are provided, they are unioned
into a single sf object using \code{\link[sf:geos_combine]{sf::st_union()}}}

\item{crop}{If \code{TRUE}, x is cropped to y using \code{\link[sf:st_crop]{sf::st_crop()}}.}

\item{trim}{If \code{TRUE}, x is trimmed to y with \code{\link[sfext:st_trim]{st_trim()}}.}

\item{collapse_types}{If \code{TRUE} (default), create new ownership type and
transaction type columns that collapse less frequent categories into a
single combined other category.}

\item{...}{Additional parameters passed to \code{\link[getdata:get_location_data]{getdata::get_location_data()}}}

\item{crs}{Coordinate reference system to return.}
}
\value{
A sf data frame.
}
\description{
\code{\link[=get_mls_data]{get_mls_data()}} is designed to use data from an internal data file provided
to the Baltimore City Department of Planning by Live Baltimore for analysis
of property sales data.
}
