% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_mls_data}
\alias{label_mls_data}
\title{Label MLS data using \code{mls_dictionary} and \code{\link[getdata:format_data]{getdata::label_with_xwalk()}}}
\usage{
label_mls_data(data, cols = c("label", "nm"))
}
\arguments{
\item{cols}{Column names, Default: c("label", "nm")}
}
\description{
Label MLS data using \code{mls_dictionary} and \code{\link[getdata:format_data]{getdata::label_with_xwalk()}}
}
\seealso{
\code{\link[getdata:format_data]{getdata::format_data()}}, \code{\link[getdata:make_xwalk_list]{getdata::make_xwalk_list()}}
}
