% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_property_summary}
\alias{baltimore_property_summary}
\title{Baltimore Properties by Neighborhood}
\format{
A data frame with 278 rows and 7 variables:
\describe{
\item{\code{type}}{Neighborhood (2010) type}
\item{\code{name}}{County, neighborhood (2010), or City Council District name}
\item{\code{num_properties}}{Total number of properties}
\item{\code{num_residential}}{Total number of properties with usecode attribute of R}
\item{\code{perc_residential}}{num_residential as a percent of total properties}
\item{\code{median_land_value}}{Median value for currland attribute (removing NA values and replacing 0 with NA)}
\item{\code{median_imprv_value}}{Median value for currimpr attribute (removing NA values and replacing 0 with NA)}
\item{\code{county}}{County name}
\item{\code{state}}{State name}
\item{\code{geography}}{Geography name}
}
}
\usage{
baltimore_property_summary
}
\description{
A summary of July 2023 real property data downloaded from Baltimore City with
\code{\link[mapbaltimore:cache_baltimore_data]{mapbaltimore::cache_baltimore_property()}} by neighborhood and City Council
district.
}
\keyword{datasets}
