% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mls_data.R
\name{format_mls_data}
\alias{format_mls_data}
\alias{format_mls_dates}
\alias{format_mls_prices}
\alias{format_mls_price_per_area}
\alias{collapse_mls_transaction_type}
\alias{collapse_mls_ownership_type}
\title{Format MLS data}
\usage{
format_mls_data(data, quarter_type = "date_last", ...)

format_mls_dates(
  data,
  list_date_col = "listing_date",
  close_date_col = "close_date",
  contract_date_col = "contract_date",
  orders = "ymd HMS",
  tz = "UTC",
  quarter_type = "date_last",
  dom_col = "dom",
  cdom_col = "cdom"
)

format_mls_prices(
  data,
  original_list_price_col = "original_list_price",
  sale_price_col = "sales_price",
  ...
)

format_mls_price_per_area(
  data,
  sale_price_col = "sales_price",
  price_per_area_unit_col = "per_sq_ft_price",
  range = c(0, 250),
  na_values = c(0, 1),
  area_col = "interior_sq_ft"
)

collapse_mls_transaction_type(
  data,
  transaction_type_col = "transaction_type",
  type_col = "sale_type"
)

collapse_mls_ownership_type(
  data,
  ownership_type_col = "ownership_type",
  collapsed_col = "owner_type",
  other_level = "Other (Condo, Coop, etc.)"
)
}
\arguments{
\item{data}{A data frame of MLS data to format.}

\item{list_date_col}{List date column name, Default: 'listing_date'}

\item{close_date_col}{PARAM_DESCRIPTION, Default: 'close_date'}

\item{contract_date_col}{Contract date column name, Default: 'contract_date'}

\item{orders}{a character vector of date-time formats. Each order string is
a series of formatting characters as listed in \code{\link[base:strptime]{base::strptime()}} but
might not include the \code{"\%"} prefix. For example, "ymd" will match all the
possible dates in year, month, day order. Formatting orders might include
arbitrary separators. These are discarded. See details for the implemented
formats. If multiple order strings are supplied, they are applied in turn
for \code{parse_date_time2()} and \code{fast_strptime()}. For \code{parse_date_time()}
the order of applied formats is determined by \code{select_formats} parameter.}

\item{tz}{a character string that specifies the time zone with which to
parse the dates}

\item{dom_col, cdom_col}{Days on market and cumulative days on market columns.
Defaults to "dom" and "cdom".}

\item{original_list_price_col, sale_price_col}{Column names for the original
list price and sales price. Defaults to "original_list_price" and
"sales_price".}

\item{price_per_area_unit_col}{Column name to add for derived price per area
unit column (based on supplied \code{area_col} parameter).}

\item{with_year}{Passed to \code{\link[lubridate:quarter]{lubridate::quarter()}} Default: \code{TRUE}}
}
\value{
A data frame.
}
\description{
\code{\link[=format_mls_data]{format_mls_data()}} calls \code{\link[=format_mls_prices]{format_mls_prices()}},
\code{\link[=format_mls_price_per_area]{format_mls_price_per_area()}}, and \code{\link[=format_mls_dates]{format_mls_dates()}}

Both \code{\link[=collapse_mls_ownership_type]{collapse_mls_ownership_type()}} and \code{\link[=collapse_mls_transaction_type]{collapse_mls_transaction_type()}}
are called by \code{\link[=get_mls_data]{get_mls_data()}} when \code{collapse_types = TRUE}
}
\seealso{
\code{\link[naniar:replace_with_na]{naniar::replace_with_na()}}
\code{\link[dplyr:mutate]{dplyr::mutate()}}, \code{\link[dplyr:distinct]{dplyr::distinct()}}
}
